Fix vki-netbsd.h to make vki_rlimit definition match netbsd's Fix m_main.c to make setup_file_descriptors work right
git-svn-id: http://svn.berlios.de/svnroot/repos/vg4nbsd/branches/aspacem@108 553a31c3-2ffc-0310-903e-cdb4b4d06021

+++ valgrind/include/vki-netbsd.h
@@ -769,8 +769,8 @@ struct	vki_rusage {
 };
 
 struct vki_rlimit {
-	unsigned long	rlim_cur;
-	unsigned long	rlim_max;
+	long long int rlim_cur;
+	long long int	rlim_max;
 };
 
 //----------------------------------------------------------------------
